rust: suggestions from clippy.

+++ src/libs/rust/resmng/src/requests.rs
@@ -116,7 +116,7 @@ fn handle_request_async(mut is: GateIStream) {
             log!(crate::LOG_DEF, "{}: {:?} failed: {}", child.name(), op, e);
             reply_vmsg!(is, e.code() as u64)
         },
-        Ok(_) => reply_vmsg!(is, 0 as u64),
+        Ok(_) => reply_vmsg!(is, 0u64),
     }
     .ok(); // ignore errors; we might have removed the child in the meantime
 }
@@ -190,7 +190,7 @@ fn alloc_pe(is: &mut GateIStream, child: &mut dyn Child) -> Result<(), Error> {
 
     child
         .alloc_pe(dst_sel, desc)
-        .and_then(|(id, desc)| reply_vmsg!(is, 0 as u64, id, desc.value()))
+        .and_then(|(id, desc)| reply_vmsg!(is, 0u64, id, desc.value()))
 }
 
 fn free_pe(is: &mut GateIStream, child: &mut dyn Child) -> Result<(), Error> {
